Add support for other runlength encoded files
This includes local radar images. DataZ and AtZ() is introduced for 3D CAPPI
products.  Data and DataZ are both slices of PlainData, stripping elevation
views and splitting concatenated layers.
A lookup catalog is neccessary, as runlength decoding requires dimension
information, that is not specified in the local picture product headers.